requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.11
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.0
watchdog==6.0.0
